Documentation

GalerkinEtena.jl

Documentation for GalerkinEtena.jl

GalerkinEtena.DGDiscretization — Method.

[x, vmapM, vmapP] = DGDiscretization(m::Mesh1D, ξ::Array{Float64,1})

return a DG discretization along the mesh and the local discretization ξ

source

GalerkinEtena.GLT — Method.

GLT(N) computes the Gauß-Lobatto-Чебышёв points defined by

\[T_k = -\cos \left( \frac{k-1}{n-1}\pi \right)\]

source

GalerkinEtena.JacobiGL — Method.

JacobiGL(α::Float64, β::Float64, N::Int)

Compute the N'th order Gauß Lobatto quadrature formula points

source

GalerkinEtena.JacobiGQ — Method.

JacobiGQ(α::Float64,β::Float64,N::Int)

return the N'th Gauß quadrature points

source

GalerkinEtena.JacobiP — Method.

JacobiP evaluate the Jacobi polynomial of type (α,β) > -1 (α+β ≢ -1) at points x for order N Note : the Jacobi polynomial is normalize by a factor γₙ = √(2/(2n+1)) taken from nodal-dg matlab code [https://github.com/tcew/nodal-dg]

source

GalerkinEtena.Legendre — Method.

Legendre(x::Array{Float64}, n::Int) computes the matrices

\[P_{ij} = P^j(x_i)\]

and

\[P'_{ij} = \frac{dP^j}{dx}(x_i)\]

where

\[P^n(x) = \frac{1}{2^nn!}\frac{d^n}{dx^n}\left((x^2-1)^n\right)\]

Note

source

GalerkinEtena.computeElementaryMatrices — Method.

computeElementaryMatrices(ξ::Array{Float64,1}, Np::Int) computes the elementary matrices 𝓥, 𝓓ᵣ on the Gauß-Lobatto grid on [-1,1]

source

GalerkinEtena.computeMask — Method.

computeMask(ξ::Array{Float64,1}) retrieves the index of the boundary nodes on the reference element

source

GalerkinEtena.genGrid — Method.

genGrid(m::Mesh1D, ξ::Array{Float64}) generate a matrix corresponding to all degree of freedom points : K x #ξ

source

GalerkinEtena.integrate — Method.

integrate(f,q)

computes the numerical approximation of the integral of f on the interval [-1,1] using q as a Quadrature formula

source

GalerkinEtena.vander — Function.

vander(x, n) computes the Vandermonde matrix V(x₁, x₂, ... , xₚ) defined by

\[V_{ij} = [x_j^{i-1}]_{ij} \, \forall \, i=1,\cdots,p \, j = 1,\cdots,\n\]

optional argument n enable us to choose the number of rows

source