GalerkinEtena.jl

Documentation for GalerkinEtena.jl

GalerkinEtena.Advec1D — Type

structure for modelizing the 1D advection equation

\[\frac{\partialu}{\partial t} + a \frac{\partial}{\partial x} = 0\]

with

\[u(x,0) = \sin (x) \\ u(0,t) = -\sin (a t)\]

GalerkinEtena.Maxwell1D — Type

structure for modelizing the 1D Maxwell's equation

\[\varepsilon(x)\frac{\partialE}{\partial t} = - \frac{\partialH}{\partial x} \mbox{ and } \mu(x)\frac{\partialH}{\partial t} = - \frac{\partialE}{\partial x} \]

GalerkinEtena.DGDiscretization — Method

[x, vmapM, vmapP] = DGDiscretization(m::Mesh1D, ξ::Array{Float64,1})

return a DG discretization along the mesh and the local discretization ξ x is of size (#ξ, K ) vmapP and vmapP is a vector of size 2*K, if we reshape into a (2,K), matrix, then [vmapM[1, i], vmapM[2, i]] are the indices of internal vertices of the i element [vmapP[1, i], vmapP[2, i]] are the indices of external vertices of the i element

GalerkinEtena.JacobiGL — Method

JacobiGL(α::Float64, β::Float64, N::Int)

Compute the N'th order Gauß Lobatto quadrature formula points

GalerkinEtena.JacobiGQ — Method

JacobiGQ(α::Float64,β::Float64,N::Int)

return the N'th Gauß quadrature points

GalerkinEtena.JacobiP — Method

JacobiP(x::Array{Float64,1},α::Float64,β::Float64,N::Int)

evaluates the Jacobi polynomial of type (α,β) > -1 (α+β ≢ -1) at points x for order N Note : the Jacobi polynomial is normalize by a factor γₙ = √(2/(2n+1)) adapted from nodal-dg matlab code [https://github.com/tcew/nodal-dg]

GalerkinEtena.Legendre — Function

Legendre(n::Int64, x::Array{Float64}, derive::Bool=true) computes the matrices

\[P_{ij} = P^j(x_i)\]

and

\[P'_{ij} = \frac{dP^j}{dx}(x_i)\]

where

\[P^n(x) = \frac{1}{2^nn!}\frac{d^n}{dx^n}\left((x^2-1)^n\right)\]

Note

GalerkinEtena.WarpFactor — Method

WarpFactor(N::Int64, rout::Array{Float64,1}) compute the warping function as defined p. 177 in Warburton-Hesthaven

GalerkinEtena.computeElementaryMatrices — Method

computeElementaryMatrices(ξ::Array{Float64,1}, Np::Int) computes the elementary matrices 𝓥, 𝓓ᵣ on the Gauß-Lobatto grid on [-1,1]

GalerkinEtena.computeMask — Method

computeMask(ξ::Array{Float64,1}) retrieves the index of the boundary nodes on the reference element

GalerkinEtena.genGrid — Method

genGrid(m::Mesh1D, ξ::Array{Float64}) generate a (#ξ,K) matrix corresponding to all degree of freedom points :

\[G_{ij} = x^k_i \in D^k\]

GalerkinEtena.integrate — Method

integrate(f,q)

computes the numerical approximation of the integral of f on the interval [-1,1] using q as a Quadrature formula

GalerkinEtena.nodes2D — Method

nodes2D(N::Int64) computes the interpolation nodes (x,y) nodes in equilateral triangle for polynomial of order N

GalerkinEtena.npToN — Method

npToN(np::Int64)

GalerkinEtena.rhs1D — Method

compute the right hand side of the advection problem

\[\frac{du_h^k}{dt} = -a \mathcal{D}_r u^_h^k-(\mathcal{M}^k)^{-1} + ...\]

GalerkinEtena.rhs1D — Method

compute the right hand side of the maxwell equation u = [E, H]

GalerkinEtena.rsToAb — Method

rsToAb(r::Array{Float64,1}, s::Array{Float64,1}) changes (r,s) coordinates to (a,b) coordinates

GalerkinEtena.xyToRs — Method

function xyToRs(x::Array{Float64,1}, y::Array{Float64,1}) convert (x,y) coords in equilateral triangle to (r,s) coordinates standard triangle I = [(-1,-1), (1,-1), (-1,1)]

GalerkinEtena.𝓔 — Method

computes flux integral